requirements.txt
django
djangorestframework
pillow
pytesseract
